pyproject.toml
[project]
name = "gnss-tools"
version = "0.1.0"
description = "Collect GNSS satellite data & plot heatmap."
readme = "README.md"
authors = [
    { name = "Curtis Doty", email = "Curtis@GreenKey.net" }
]

# only tested on 3.13 so far
requires-python = ">=3.13"

dependencies = [
  "gpsd==3.26.*",
  "matplotlib==3.10.*",
  "numpy==2.3.*",
  "pandas==2.3.*",
]

[tool.pytest.ini_options]
addopts = "--cov=skylog --cov=gnss"
pythonpath = [
  "."
]
